refactor(tui): extract the help strip into tui::render::footer
Footer segment tables, the width ladder, chord badges, and the copy and
notice flashes move together with their tests. Shared layout fixtures
move to tui::test_support so submodule tests can build a snapshot.
